+++ lib/db.dart
@@ -4,12 +4,12 @@ FutureOr db(List<String> args) async {
   //
   // Create cli commands for database:
   //
-  // lucy db init
-  // lucy db create
-  // lucy db model --name User --attributes
-  // lucy db migrate
-  // lucy db seed --name users
-  // lucy db seeds
+  // l db init
+  // l db create
+  // l db model --name User --attributes
+  // l db migrate
+  // l db seed --name users
+  // l db seeds
   //
   print('Command db is under development...');
 }