+++ lib/lucy.dart
@@ -2,6 +2,7 @@ export 'build.dart';
 export 'controller.dart';
 export 'create.dart';
 export 'db.dart';
+export 'middleware.dart';
 export 'repository.dart';
 export 'run.dart';
 export 'version.dart';